requirements.txt
streamlit>=1.28.0
streamlit-webrtc>=0.45.0
ultralytics>=8.0.0
numpy<2.0
opencv-python-headless<=4.10.0.84
Pillow>=10.0.0
av>=10.0.0